refactor: redistribute 48 tests from main.rs to owning modules
Move 48 #[cfg(test)] unit tests from src/main.rs to their owning modules:
- src/cli.rs: 10 tests (CLI completion + porcelain parser)
- src/domain/worktree.rs: 24 tests (display path, parsers, table formatter)
- src/commands/common.rs: 14 tests (worktree resolution, path canonicalization)

Production behavior is unchanged. The src/main.rs file shrinks from 919
to 66 lines (production only). Total cargo test count increases from 553
to 601 because the moved tests now compile into both the lib and bin
targets via src/lib.rs's `pub mod` declarations.

One test (test_display_path_outside_home) was renamed to
test_display_path_outside_home_unix_exact to avoid a name collision with
a pre-existing platform-aware test of the same name in domain/worktree.rs.
